+++ Source/Core/HCAssertThat.m
@@ -14,6 +14,7 @@
 #import "HCTestFailure.h"
 #import "HCTestFailureHandler.h"
 #import "HCTestFailureHandlerChain.h"
+#import <libkern/OSAtomic.h>
 
 
 static NSString *describeMismatch(id matcher, id actual)
@@ -26,17 +27,42 @@ static NSString *describeMismatch(id matcher, id actual)
     return [description description];
 }
 
+static void reportMismatch(id testCase, id actual, id <HCMatcher> matcher,
+                           char const *fileName, int lineNumber)
+{
+    HCTestFailure *failure = [[HCTestFailure alloc] initWithTestCase:testCase
+                                                            fileName:[NSString stringWithUTF8String:fileName]
+                                                          lineNumber:(NSUInteger)lineNumber
+                                                              reason:describeMismatch(matcher, actual)];
+    HCTestFailureHandler *chain = HC_testFailureHandlerChain();
+    [chain handleFailure:failure];
+}
+
 void HC_assertThatWithLocation(id testCase, id actual, id <HCMatcher> matcher,
                                const char *fileName, int lineNumber)
 {
     if (![matcher matches:actual])
-    {
-        HCTestFailure *failure = [[HCTestFailure alloc] initWithTestCase:testCase
-                                                                fileName:[NSString stringWithUTF8String:fileName]
-                                                              lineNumber:(NSUInteger)lineNumber
-                                                                  reason:describeMismatch(matcher, actual)];
-        HCTestFailureHandler *chain = HC_testFailureHandlerChain();
-        [chain handleFailure:failure];
-    }
+        reportMismatch(testCase, actual, matcher, fileName, lineNumber);
 }
 
+void HC_assertThatAfterWithLocation(id testCase, NSTimeInterval maxTime,
+                                    HCAssertThatAfterActualBlock actualBlock, id<HCMatcher> matcher,
+                                    const char *fileName, int lineNumber)
+{
+    BOOL match = NO;
+    id actual = nil;
+    NSTimeInterval timeOut = maxTime;
+    NSDate *expiryDate = [NSDate dateWithTimeIntervalSinceNow:timeOut];
+    while (1)
+    {
+        actual = actualBlock();
+        match = [matcher matches:actual];
+        if (match || ([[NSDate date] compare:expiryDate] == NSOrderedDescending))
+            break;
+        [[NSRunLoop currentRunLoop] runUntilDate:[NSDate dateWithTimeIntervalSinceNow:0.01]];
+        OSMemoryBarrier();
+    }
+
+    if (!match)
+        reportMismatch(testCase, actual, matcher, fileName, lineNumber);
+}