+++ CHANGES.md
@@ -5,9 +5,11 @@ NEXT VERSION
 
 - Fix crash when OCHamcrest tries to describe an OCMockito mock object. _Thanks to: Michael Seghers_
 
-**New matchers:**
+**Features:**
 
-- `everyItem` matches collections if every item satisfies a given matcher.
+- `assertThatAfter` tests asynchronous code, retrying the assertion until a given timeout.
+  Wrap the code you want to evaluate in `futureValueOf`. _Thanks to: Sergio Padrino_
+- New matcher `everyItem` matches collections if every item satisfies a given matcher.
 
 **Improvements:**